chore() extract more images pill component
