+++ ui/design/src/components/ImageGallery/entry-image-gallery.tsx
@@ -1,15 +1,7 @@
 import * as React from 'react';
-import styled from 'styled-components';
-import { Box, Text } from 'grommet';
+import { MoreImagesPill } from './more-images-pill';
 import { ImageGridItem, ImageObject } from './image-grid-item';
 import { SingleImageContainer, StyledGrid } from './index';
-
-const StyledPill = styled(Box)`
-  position: absolute;
-  bottom: 21px;
-  right: 16px;
-  z-index: 1;
-`;
 
 export interface IImageGallery {
   imagesLabel?: string;
@@ -26,24 +18,7 @@ const EntryImageGallery: React.FC<IImageGallery> = props => {
     return images?.length === 1;
   }, [images]);
 
-  const MoreImagesPill = () => {
-    const hiddenImages = images.length - 4;
-    const label = hiddenImages > 1 ? imagesLabel : imageLabel;
-    return (
-      <StyledPill
-        direction="row"
-        round="large"
-        background={{ color: 'accent' }}
-        fill="horizontal"
-        width={{ max: '6rem' }}
-        height="1.5rem"
-        align="center"
-        justify="center"
-      >
-        <Text color="white">{`+${hiddenImages} ${label}`}</Text>
-      </StyledPill>
-    );
-  };
+  const hiddenImagesNumber = React.useMemo(() => images.length - 4, [images]);
 
   // limit shown images to 4
   const displayedImages = React.useMemo(() => {
@@ -76,14 +51,15 @@ const EntryImageGallery: React.FC<IImageGallery> = props => {
           images={images}
         />
       ))}
-      {images.length > 4 && <MoreImagesPill />}
+      {images.length > 4 && (
+        <MoreImagesPill
+          hiddenImages={hiddenImagesNumber}
+          imageLabel={imageLabel}
+          imagesLabel={imagesLabel}
+        />
+      )}
     </StyledGrid>
   );
-};
-
-EntryImageGallery.defaultProps = {
-  imagesLabel: 'images',
-  imageLabel: 'image',
 };
 
 export { EntryImageGallery };
